Dropped weak reasoning on the constraints to be included
If there was only a noinherit check on the parent table the child would
have receive no check so it would have created the one explicitely
marked as noinherit, which sounds wrong.

We can't either look at the noinherit field on the constraint to decide
what to copy and what not as all the fkey/pkey etc. have noinherit=true.

So let's hardcode the info that subclassing copies all and only the
checks. Added a parameter to the more generic function to get the
constraint types to avoid copying.